+++ mypyc/irbuild/mapper.py
